requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0.post0
ta==0.11.0

# Messaging
python-telegram-bot==21.4

# Optional (when you wire a real broker / DB)
redis==5.0.8
psycopg2-binary==2.9.9

# Testing
pytest==8.3.2
